update from Recras/sf1 branch php81-compat

+++ lib/cache/sfFileCache.class.php
@@ -156,7 +156,7 @@ class sfFileCache extends sfCache
     $result = true;
     foreach (new RecursiveIteratorIterator(new RecursiveDirectoryIterator($this->getOption('cache_dir'))) as $file)
     {
-      if (sfCache::ALL == $mode || !$this->isValid($file))
+      if ((sfCache::ALL == $mode || !$this->isValid($file)) && !is_dir($file))
       {
         $result = @unlink($file) && $result;
       }
@@ -255,8 +255,10 @@ class sfFileCache extends sfCache
       {
         fseek($fp, 0, SEEK_END);
         $length = ftell($fp) - 24;
-        fseek($fp, 24);
-        $data[self::READ_DATA] = @fread($fp, $length);
+        if ($length >= 0) {
+          fseek($fp, 24);
+          $data[self::READ_DATA] = @fread($fp, $length);
+        }
       }
     }
     else
@@ -311,7 +313,7 @@ class sfFileCache extends sfCache
     {
       if (copy($tmpFile, $path))
       {
-        unlink($tmpFile);
+        sfToolkit::safeUnlink($tmpFile);
       }
     }
